backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu
scikit-learn
numpy
pandas
